changed directory system to help with making a library. moved the SimInfo class to its own file in the root directory. Changed the body directory to the structure directory. Created a util directory and moved the neighborgrid code there. Added some init files and a setup file to make a python library out of the modules. Have tested that this works. Tests still pass with the new framework.
