+++ setup.py
@@ -1,0 +1,11 @@
+from setuptools import find_packages, setup
+
+setup(
+    name='SAASH',
+    packages=find_packages(),
+    version='0.1.0',
+    description='Self-Assembly Analysis Script for HOOMD',
+    author='onehalfatsquared',
+    license='MIT',
+    test_suite='testing',
+)
