+++ testing/patchy_2d/interactions.txt
@@ -1,0 +1,2 @@
+P P 0.06
+S 1
