+++ testing/unit/clustering_tests.py
@@ -3,14 +3,14 @@ import pytest
 
 #add the paths to the source code folder for testing
 import sys
-sys.path.insert(0, '../../src')
+sys.path.insert(0, '../../SAASH')
 
 import numpy as np
 
-from body import body
-from body import neighborgrid
-from body import cluster
-from body import frame
+from structure import body
+from structure import cluster
+from structure import frame
+from util import neighborgrid
 
 import clustering_examples as examples
